# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AttemptOut": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/MetaCandidateOut"
            },
            "title": "Candidates",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "candidates"
        ],
        "title": "AttemptOut",
        "type": "object"
      },
      "BrandingAssetOut": {
        "properties": {
          "asset": {
            "title": "Asset",
            "type": "string"
          },
          "preview_url": {
            "title": "Preview Url",
            "type": "string"
          }
        },
        "required": [
          "asset",
          "preview_url"
        ],
        "title": "BrandingAssetOut",
        "type": "object"
      },
      "BrandingDefaultsIn": {
        "properties": {
          "caption_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption Template"
          },
          "intro_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intro Asset"
          },
          "intro_duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intro Duration"
          },
          "logo_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Asset"
          },
          "outro_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outro Asset"
          },
          "outro_duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outro Duration"
          }
        },
        "title": "BrandingDefaultsIn",
        "type": "object"
      },
      "BrandingDefaultsOut": {
        "properties": {
          "caption_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption Template"
          },
          "intro_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intro Asset"
          },
          "intro_duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intro Duration"
          },
          "logo_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Asset"
          },
          "outro_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outro Asset"
          },
          "outro_duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outro Duration"
          }
        },
        "required": [
          "logo_asset",
          "intro_asset",
          "intro_duration",
          "outro_asset",
          "outro_duration",
          "caption_template"
        ],
        "title": "BrandingDefaultsOut",
        "type": "object"
      },
      "BrandingIn": {
        "properties": {
          "branding": {
            "additionalProperties": true,
            "title": "Branding",
            "type": "object"
          }
        },
        "required": [
          "branding"
        ],
        "title": "BrandingIn",
        "type": "object"
      },
      "BrandingPatchIn": {
        "additionalProperties": false,
        "properties": {
          "caption_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption Template"
          },
          "intro_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intro Asset"
          },
          "intro_duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intro Duration"
          },
          "logo_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Asset"
          },
          "outro_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outro Asset"
          },
          "outro_duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outro Duration"
          }
        },
        "title": "BrandingPatchIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MetaCandidateOut": {
        "properties": {
          "ig_user_id": {
            "title": "Ig User Id",
            "type": "string"
          },
          "ig_username": {
            "title": "Ig Username",
            "type": "string"
          },
          "page_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Id"
          },
          "page_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Name"
          }
        },
        "required": [
          "ig_user_id",
          "ig_username"
        ],
        "title": "MetaCandidateOut",
        "type": "object"
      },
...
      "SelectIn": {
        "properties": {
          "ig_user_id": {
            "title": "Ig User Id",
            "type": "string"
          }
        },
        "required": [
          "ig_user_id"
        ],
        "title": "SelectIn",
        "type": "object"
      },
...
      "StartIn": {
        "properties": {
          "return_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Uri"
          }
        },
        "title": "StartIn",
        "type": "object"
      },
      "StartOut": {
        "properties": {
          "attempt_id": {
            "title": "Attempt Id",
            "type": "string"
          },
          "auth_url": {
            "title": "Auth Url",
            "type": "string"
          }
        },
        "required": [
          "auth_url",
          "attempt_id"
        ],
        "title": "StartOut",
        "type": "object"
      },
      "StatusOut": {
        "properties": {
          "connection_type": {
            "default": "facebook_login",
            "title": "Connection Type",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "health": {
            "title": "Health",
            "type": "string"
          },
          "ig_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ig User Id"
          },
          "ig_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ig Username"
          },
          "last_checked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checked At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_refreshed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Refreshed At"
          },
          "page_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Id"
          },
          "page_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Name"
          }
        },
        "required": [
          "health"
        ],
        "title": "StatusOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Dojo publishing API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/meta/instagram/token": {
      "post": {
        "operationId": "connect_instagram_token_api_meta_instagram_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Connect Instagram Token",
        "tags": [
          "meta"
        ]
      }
    },
    "/api/meta/oauth/attempts/{attempt_id}": {
      "get": {
        "operationId": "get_attempt_api_meta_oauth_attempts__attempt_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Attempt",
        "tags": [
          "meta"
        ]
      }
    },
    "/api/meta/oauth/attempts/{attempt_id}/select": {
      "post": {
        "operationId": "select_account_api_meta_oauth_attempts__attempt_id__select_post",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Account",
        "tags": [
          "meta"
        ]
      }
    },
    "/api/meta/oauth/callback": {
      "get": {
        "operationId": "oauth_callback_api_meta_oauth_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "title": "State",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Oauth Callback Api Meta Oauth Callback Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Oauth Callback",
        "tags": [
          "meta"
        ]
      }
    },
    "/api/meta/oauth/start": {
      "post": {
        "operationId": "oauth_start_api_meta_oauth_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Oauth Start",
        "tags": [
          "meta"
        ]
      }
    },
...
    "/api/packages/active/branding": {
      "get": {
        "operationId": "get_branding_api_packages_active_branding_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Branding Api Packages Active Branding Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Branding",
        "tags": [
          "packages"
        ]
      },
      "put": {
        "operationId": "set_branding_api_packages_active_branding_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandingIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Set Branding Api Packages Active Branding Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Branding",
        "tags": [
          "packages"
        ]
      }
    },
...
    "/api/settings/branding": {
      "get": {
        "operationId": "get_branding_defaults_api_settings_branding_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandingDefaultsOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Branding Defaults",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "patch_branding_defaults_api_settings_branding_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandingPatchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandingDefaultsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Branding Defaults",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "set_branding_defaults_api_settings_branding_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandingDefaultsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandingDefaultsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Branding Defaults",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/branding/assets": {
      "post": {
        "operationId": "upload_branding_image_api_settings_branding_assets_post",
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandingAssetOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Upload Branding Image",
        "tags": [
          "branding assets"
        ]
      }
    },
    "/api/settings/branding/assets/{asset_id}": {
      "get": {
        "operationId": "preview_branding_image_api_settings_branding_assets__asset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Branding Image",
        "tags": [
          "branding assets"
        ]
      }
    },
...
  }
}